feat: Persist chat history at lifecycle boundaries and add transcript scrollbar
Persist conversations only at explicit boundaries such as question submission, response completion or cancellation, completed Git actions, and conversation transitions; streaming deltas now update memory only. Interrupted responses keep received text with an interruption notice instead of being removed. Add the shared sidebar overlay scroll indicator bridge to the transcript and render loaded messages eagerly to improve long Markdown measurement.
