+++ macgit/Views/MainWindow/ScrollViewIndicatorController.swift
@@ -1,0 +1,40 @@
+//
+//  macgit (Commit+) - a macOS Git client built with Swift and SwiftUI.
+//  Copyright (C) 2026  Thanh Tran <trantienthanh2412@gmail.com>
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+import AppKit
+import SwiftUI
+
+struct ScrollViewIndicatorController: NSViewRepresentable {
+    let showsIndicators: Bool
+    var controlSize: NSControl.ControlSize? = nil
+    let onScroll: () -> Void
+
+    func makeNSView(context: Context) -> ScrollViewIndicatorBridgeView {
+        let view = ScrollViewIndicatorBridgeView()
+        view.configure(showsIndicators: showsIndicators, controlSize: controlSize, onScroll: onScroll)
+        return view
+    }
+
+    func updateNSView(_ nsView: ScrollViewIndicatorBridgeView, context: Context) {
+        nsView.configure(showsIndicators: showsIndicators, controlSize: controlSize, onScroll: onScroll)
+    }
+
+    static func dismantleNSView(_ nsView: ScrollViewIndicatorBridgeView, coordinator: ()) {
+        nsView.detach()
+    }
+}
